Refactoring DownloadData
As I looked into DownloadData, I found it to be very redundant:
There are distinct but similar methods to open, getSize, etc,  all 3 types
of files (apk, expansion and split).

This is exactly where a class makes sense, so I've created `AppFile` class,
with can proble with getSize() and openStream().

It's subclasses add useful methods if necessary, like `SplitApkFile.getId()`
and `AdditionalFile.getVersionCode()`.

DownloadData can now be resumed to `getMainApk()`, `getAdditionalFiles()` and `getSplitApkFiles()`,
all of which acts exactly the same.

This new structure makes the code for `AppDownloadWorker` and `cli.Play.downloadApp` considerably simpler.