Improve GooglePlayException: Preserve HTTP status and extract pretty messages from response blobs
The HTTP status is useful for handling errors in my code -- E.g., I was getting 429 errors, and now I handle it as sleep-and-retry.

The message string used to be the raw response buffer. While it was usable, the protobuf bytes and HTML tags make it noisy.

To handle that, I added a very simple regex to remove HTML tags, and a small function extracts ascii strings from protobuf blobs. If none of the strategies work,
it falls back to the showing the HTTP status line.

While it is a simple (And a bit crappy), I've found them to work well enough:
- HTML error: `<html><body><h1>429 Too Many Requests</h1><p>Please reduce your request rate.</p></body></html>` -> `429 Too Many Requests - Please reduce your request rate.`
- Protobuf error: `????Item not found.`  =>  `Item not found.`